requirements.txt
zxcvbn-python=4.4.16